Remove Notification repo and notification type tbl
