--- ECommerce.Application/Repositories/Notification/Enums/NotificationType.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ECommerce.Application.Repositories.Notification.Enums
-{
-    public enum NotificationType
-    {
-        Other = 1,
-        OrderStatus = 2,
-        Comment = 3,
-        Coupon = 4,
-        DiscountFromShop = 5,
-        NewProduct = 6,
-        Like = 7,
-        UpdateComment = 8,
-        NewMessage = 9
-    }
-}
