+++ ECommerce.Application/Services/Notifications/INotificationService.cs
@@ -1,21 +1,21 @@
 ﻿using ECommerce.Application.Common;
-using ECommerce.Application.Repositories;
-using ECommerce.Application.Repositories.Notification;
-using ECommerce.Application.Repositories.Notification.Dtos;
+using ECommerce.Data.Entities.ProductSchema;
+using ECommerce.Data.Entities.UserSchema;
+using ECommerce.Dtos.UserSchema;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ECommerce.Application.Services.Notification
+namespace ECommerce.Application.Services.Notifications
 {
     public interface INotificationService
     {
-        INotificationRepository Notification { get; }
         Task<Response<NotificationModel>> ReadAsync(int id = 0);
         Task<Response<string>> DeleteAsync(int id = 0);
         Task<List<NotificationModel>> GetAllByUserIdAsync(int userId = 0);
-        Task SaveAsync();
+        Task<Notification> CreateCommentNotiAsync(Rate comment);
+        Task<Notification> CreateLikeDislikeNotiAsync(Rate comment);
     }
 }
