--- ECommerce.Data/Entities/UserSchema/NotificationType.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-#nullable disable
-
-namespace ECommerce.Data.Entities.UserSchema
-{
-    public partial class NotificationType
-    {
-        public NotificationType()
-        {
-            Notifications = new HashSet<Notification>();
-        }
-
-        public int Id { get; set; }
-        public string TypeName { get; set; }
-        public string TypeCode { get; set; }
-
-        public virtual ICollection<Notification> Notifications { get; set; }
-    }
-}
