+++ ECommerce.Dtos/UserSchema/NotificationModel.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace ECommerce.Application.Repositories.Notification.Dtos
+namespace ECommerce.Dtos.UserSchema
 {
     public class NotificationModel
     {
