+++ ECommerce.Utilities/Constants/NotificationConstant.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ECommerce.Utilities.Constants
+{
+    public class NotificationConstant
+    {
+        public const string OTHER = "OTHER";
+        public const string ORDER_STATUS = "ORDER_STATUS";
+        public const string COMMNENT = "COMMENT";
+        public const string COUPON = "COUPON";
+        public const string DISCOUNT_FROM_SHOP = "DISCOUNT_FROM_SHOP";
+        public const string NEW_PRODUCT = "NEW_PRODUCT";
+        public const string LIKE = "LIKE";
+        public const string UPDATE_COMMENT = "UPDATE-COMMENT";
+        public const string NEW_MESSAGE = "NEW_MESSAGE";
+    }
+}
